# docs/openapi/platform-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Observability Platform API",
    "description": "File n�y du?c g?p th? c�ng t? /v3/api-docs c?a ingestion-service v� core-app t?i th?i di?m t?o � c� th? l?ch n?u code sau n�y d?i m� kh�ng regenerate l?i; ngu?n s? th?t lu�n l� 2 endpoint /v3/api-docs s?ng.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/alerts/{id}/resolve": {
      "patch": {
        "tags": [
          "Alerts"
        ],
        "summary": "Resolve an alert",
        "description": "Changes the status of a TRIGGERED or ACKNOWLEDGED alert to RESOLVED. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "resolve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert resolved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseAlertResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Alert not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (alert already resolved or cannot be resolved)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/alerts/{id}/acknowledge": {
      "patch": {
        "tags": [
          "Alerts"
        ],
        "summary": "Acknowledge an alert",
        "description": "Changes the status of a TRIGGERED alert to ACKNOWLEDGED. Yêu cầu role: ADMIN, DEVOPS, DEVELOPER.",
        "operationId": "acknowledge",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert acknowledged successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseAlertResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Alert not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (alert not in correct state for acknowledgment)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "List alerts",
        "description": "Returns a list of alerts filtered by service, environment, or status. Status is a single string, e.g. TRIGGERED, ACKNOWLEDGED, RESOLVED. Yêu cầu role: ADMIN, DEVOPS, DEVELOPER, VIEWER (authenticated).",
        "operationId": "list_2",
        "parameters": [
          {
            "name": "serviceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of alerts",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseListAlertResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UnifiedResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "object",
            "description": "The main response data payload"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "AlertResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ruleId": {
            "type": "integer",
            "format": "int64"
          },
          "serviceId": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "windowStart": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "triggeredAt": {
            "type": "string"
          },
          "acknowledgedAt": {
            "type": "string"
          },
          "resolvedAt": {
            "type": "string"
          },
          "occurrenceCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "The main response data payload"
      },
      "UnifiedResponseAlertResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "$ref": "#/components/schemas/AlertResponse"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "UnifiedResponseListAlertResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "array",
            "description": "The main response data payload",
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            }
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "X-API-Key": {
        "type": "apiKey",
        "description": "API Key authentication. Required for all endpoints.",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}